fixes #1005 use vite to asset bundle instead of webpack
