+++ .gitignore
@@ -52,18 +52,10 @@ nbproject/*
 .vscode
 # Sass preprocessor
 .sass-cache/
-# phpenv
+
+# anyenv
 .php-version
-# nodenv
 .node-version
-# pyenv
-.python-version
 
 # phpunit cache
 .phpunit.result.cache
-
-# Vetur
-vetur.config.js
-
-# cakephp-vite
-/webroot/build
