+++ config/app.php
@@ -425,9 +425,4 @@ return [
         'dsn' => env('SENTRY_DSN'),
         'environment' => env('CAKE_ENV', 'local'),
     ],
-
-    'ViteHelper' => [
-        'forceProductionMode' => true,
-        'mainJs' => 'main.ts',
-    ],
 ];
