+++ package.json
@@ -5,15 +5,14 @@
   "description": "gotea front development tools configuration",
   "author": "Kazuki Kamizuru",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite build --mode development --watch",
     "lint": "run-p -c lint:*",
     "lint:scripts": "eslint --ext .ts,.vue ./resources/assets/scripts",
     "lint:styles": "stylelint \"./resources/assets/styles/**/*.scss\" \"./resources/assets/scripts/**/*.vue\"",
     "format": "run-p -c format:*",
     "format:scripts": "npm run lint:scripts -- --fix",
     "format:styles": "npm run lint:styles -- --fix",
-    "prod": "vite build",
-    "webpack": "webpack --config=./webpack.config.js"
+    "prod": "vite build"
   },
   "main": "resources/scripts/main.ts",
   "dependencies": {
@@ -24,48 +23,33 @@
     "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@types/glob": "^7.2.0",
     "@types/node": "^14.18.23",
     "@types/pikaday": "^1.6.5",
     "@typescript-eslint/eslint-plugin": "^4.31.0",
     "@typescript-eslint/parser": "^4.31.0",
     "@vitejs/plugin-vue": "^3.0.1",
     "@vue/compiler-sfc": "^3.2.19",
-    "css-loader": "^6.2.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-config-typescript": "^3.0.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-vue": "^7.17.0",
-    "eslint-webpack-plugin": "^3.2.0",
-    "fibers": "^5.0.0",
-    "glob": "^7.2.0",
-    "mini-css-extract-plugin": "^2.2.2",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.14",
     "postcss-html": "^1.5.0",
     "postcss-scss": "^4.0.3",
     "prettier": "^2.3.2",
-    "sass": "^1.39.0",
-    "sass-loader": "^12.1.0",
-    "sass-resources-loader": "^2.2.4",
-    "style-loader": "^3.3.1",
+    "sass": "^1.54.1",
     "stylelint": "^14.6.0",
-    "stylelint-bare-webpack-plugin": "^2.1.0",
     "stylelint-config-recommended-scss": "^5.0.2",
     "stylelint-config-recommended-vue": "^1.3.0",
     "stylelint-config-standard": "^25.0.0",
     "stylelint-scss": "^4.3.0",
-    "ts-loader": "^9.3.1",
     "typescript": "^4.7.4",
     "vite": "^3.0.4",
-    "vue-eslint-parser": "^7.11.0",
-    "vue-loader": "16.5.0",
-    "vue-style-loader": "^4.1.3",
-    "webpack": "^5.52.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-remove-empty-scripts": "^0.8.1"
+    "vite-plugin-eslint": "^1.7.0",
+    "vite-plugin-stylelint": "^3.0.6",
+    "vue-eslint-parser": "^7.11.0"
   },
   "engines": {
     "node": ">=14.18.0 <15"
