+++ src/Application.php
@@ -77,7 +77,6 @@ class Application extends BaseApplication implements
         $this->addPlugin('Authentication');
         $this->addPlugin('Authorization');
         $this->addPlugin('Connehito/CakeSentry');
-        $this->addPlugin('ViteHelper');
     }
 
     /**
