+++ src/View/AppView.php
@@ -11,7 +11,6 @@ use Cake\View\View;
  *
  * @property \Authentication\View\Helper\IdentityHelper $Identity
  * @property \Gotea\View\Helper\DateHelper $Date
- * @property \ViteHelper\View\Helper\ViteScriptsHelper $ViteScripts
  */
 class AppView extends View
 {
@@ -23,7 +22,6 @@ class AppView extends View
         parent::initialize();
 
         $this->loadHelper('Authentication.Identity');
-        $this->loadHelper('ViteHelper.ViteScripts');
     }
 
     /**
